requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.7
google-generativeai>=0.3.0
pdfplumber>=0.10.3
python-docx>=1.1.0
fpdf2>=2.7.6
python-dotenv>=1.0.0
pydantic>=2.5.0

